requirements.txt
amqp==1.4.9
anyjson==0.3.3
appdirs==1.4.3
beautifulsoup4==4.5.3
billiard==3.3.0.23
bleach==1.4
bootstrap-admin==0.3.7.1
captcha==0.2.2
celery==3.1.25
defusedxml==0.5.0
dj-database-url==0.4.2
-e git://github.com/hersonls/djamin.git@5b5887c84341065844ab44b6cfc8ccb865ccfdb9#egg=djamin
Django==1.10.6
django-admin-bootstrapped==2.5.7
django-admin-interface==0.3.2
django-allauth==0.31.0
django-authtools==1.5.0
django-braces==1.11.0
django-celery==3.2.1
django-celery-results==1.0.1
django-colorfield==0.1.12
django-crispy-forms==1.6.1
django-debug-toolbar==1.7
django-environ==0.4.2
django-flat-responsive==1.3.0
django-flat-theme==1.1.4
django-formtools==2.0
django-geoposition==0.3.0
django-haystack==2.6.0
django-modelcluster==3.1
django-recaptcha==1.2.1
django-taggit==0.22.0
django-treebeard==4.1.0
django-uuidfield==0.5.0
django-wpadmin==1.7.4
djangorestframework==3.6.2
docker==2.2.1
docker-pycreds==0.2.1
docutils==0.13.1
easy-thumbnails==2.4.1
ez-setup==0.9
feedparser==5.1.3
gunicorn==19.7.1
html5lib==0.999
httplib2==0.9.2
Jinja2==2.9.6
kombu==3.0.37
langdetect==1.0.5
Markdown==2.4.1
markdown2==2.3.3
MarkupSafe==0.23
oauthlib==0.6.3
olefile==0.44
ordereddict==1.1
packaging==16.8
Pillow==4.0.0
psycopg2==2.7.1
Pygments==1.6
pyparsing==2.2.0
pyPdf==1.13
PyPDF2==1.26.0
python-openid==2.2.5
python3-openid==3.1.0
pytz==2017.2
pyzmail==1.0.3
reportlab==3.4.0
requests==2.13.0
requests-oauthlib==0.4.1
six==1.10.0
South==1.0.2
Sphinx==1.2.2
sqlparse==0.2.3
Unidecode==0.4.20
urllib3==1.20
vine==1.1.3
wagtail==1.9.1
webencodings==0.5.1
websocket-client==0.40.0
whitenoise==3.3.0
Whoosh==2.6.0
Willow==0.4
xadmin==0.6.1
xhtml2pdf==0.2b1
